+++ changelog.md
@@ -2,9 +2,16 @@
 Notable changes to the `cc.isr.vi` solution are documented in this file using the 
 [Keep a Changelog] style. The dates specified are in coordinated universal time (UTC).
 
-[8.1.9587]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9593]: https://www.github.com/atecoder/dn.vi.ivi
 
 Current IVI Visa Compatibility: 8.0.2.9518
+
+## [8.1.9593] - 2026-04-07
+- 2600 TTM
+  - Driver Settings
+    - Fix construction of the settings scribe.
+  - Tests
+     - Add Meter tests.
 
 ## [8.1.9587] - 2026-04-01
 - TTM